+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Fixed
 -----
 * Different line endings are correctly handled when opening gzipped files.
 * Fail gracefully if the field does not exist assertFileExists.
+* Enabled processor tests (GO, Expression, Variant Calling)
 
 Added
 -----